+++ src/Application/Exception/Speaker/EmailDoesNotExistException.php
@@ -1,0 +1,17 @@
+<?php
+
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
+namespace Proximum\Vimeet\Application\Exception\Speaker;
+
+use Throwable;
+
+class EmailDoesNotExistException extends \Exception
+{
+}