+++ tests/Application/Command/Happening/Speaker/UpdateHandlerTest.php
@@ -16,7 +16,9 @@ use Proximum\Vimeet\Application\Command\Happening\Speaker\Update;
 use Proximum\Vimeet\Application\Command\Happening\Speaker\UpdateHandler;
 use Proximum\Vimeet\Domain\Model\Happening\Speaker;
 use Proximum\Vimeet\Domain\Model\Happening\SpeakerTranslation;
+use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Repository\Happening\SpeakerRepositoryInterface;
+use Proximum\Vimeet\Domain\Repository\UserRepositoryInterface;
 use Proximum\Vimeet\Tests\Factory\EventFactory;
 
 class UpdateHandlerTest extends TestCase
@@ -26,16 +28,18 @@ class UpdateHandlerTest extends TestCase
         //Context
         $event = EventFactory::createEvent();
         $event->setLocales(['fr', 'en'], 'fr');
+        $user = $this->prophesize(User::class);
+        $user->getEmail()->shouldBeCalled()->willReturn('thomas@hotmail.com');
 
         //Expected
-        $expectedSpeaker       = new Speaker($event, 'a', 'b', 'c', '', '');
+        $expectedSpeaker       = new Speaker($event, 'a', 'b', 'c', '', '', $user->reveal());
         $expectedTranslationFR = new SpeakerTranslation($expectedSpeaker, 'fr', 'ABC');
         $expectedTranslationEN = new SpeakerTranslation($expectedSpeaker, 'en', 'DEF');
         $expectedSpeaker->getTranslations()->set('fr', $expectedTranslationFR);
         $expectedSpeaker->getTranslations()->set('en', $expectedTranslationEN);
 
         //Command
-        $speaker       = new Speaker($event, 'firstName', 'lastName', 'orga', '', '');
+        $speaker       = new Speaker($event, 'firstName', 'lastName', 'orga', '', '', $user->reveal());
         $translationFR = new SpeakerTranslation($speaker, 'fr', 'foo');
         $translationEN = new SpeakerTranslation($speaker, 'en', 'bar');
         $speaker->getTranslations()->set('fr', $translationFR);
@@ -53,15 +57,19 @@ class UpdateHandlerTest extends TestCase
                 'position' => 'DEF',
             ],
         ];
+        $update->email      = 'thomas@hotmail.com';
 
         //Mock
         $speakerRepository = $this->prophesize(SpeakerRepositoryInterface::class);
         $speakerRepository->set($expectedSpeaker)->shouldBeCalled();
 
+        $userRepository = $this->prophesize(UserRepositoryInterface::class);
+        $userRepository->findByEventAndEmail($event, 'thomas@hotmail.com')->shouldBeCalled()->willReturn($user->reveal());
+
         $fileStorage = $this->prophesize(FileStorageInterface::class);
 
         //Handler
-        $handler = new UpdateHandler($speakerRepository->reveal(), $fileStorage->reveal());
+        $handler = new UpdateHandler($speakerRepository->reveal(), $fileStorage->reveal(), $userRepository->reveal());
         $handler->handle($update);
     }
 }